Refactored home pages
