+++ lib/routes.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:portfolio_website/pages/home_page.dart';
-import 'package:portfolio_website/pages/launch_page.dart';
+import 'package:portfolio_website/pages/home/home_main.dart';
+import 'package:portfolio_website/pages/launch/launch_page.dart';
 import 'package:portfolio_website/providers/launch_page_animation_provider.dart';
 import 'package:provider/provider.dart';
 
@@ -15,7 +15,7 @@ final GoRouter routes = GoRouter(
         GoRoute(
           path: 'home',
           pageBuilder: (context, state) => CustomTransitionPage(
-            child: HomePage(),
+            child: HomeMain(),
             key: state.pageKey,
             transitionDuration: const Duration(milliseconds: 1000),
             transitionsBuilder: (_, animation, ___, child) {
